Implement #474.

The first commit makes the API change in the simplest way possible and patches dependent packages within this repository to account for the change.

Work remaining:

- [x] Update documentation.
- [x] Check for benchmark regressions.
- [x] Justify breaking change.