@willfaught 2017-03-06
>That prevents escape analysis from eliminating the allocation of the keyvals slice passed to Log.

Interesting. I'd like to understand this better. How did you know escape analysis was the culprit? Why is the slice allocated for Log on an interface instead of Log on a struct?