@ChrisHines 2017-03-06
I inferred that from a couple pieces of information.

- Looking at the generated assembly from both master and this branch it was clear that the extra allocation was happening in the benchmark code prior to calling Log on the interface value, but not on the concrete `*Context` value.
- I had a conversation with Brad Fitzpatrick at one of the recent GopherCon's in which he pointed out that one of the benefits of the stdlib Logger being a concrete type is that the compiler knows statically which methods are getting called and that this allows better escape analysis for the method arguments.