@peterbourgon 2017-03-06
This change looks good to me. I'm happy to take the breaking change.