@ChrisHines 2017-03-06
@peterbourgon What do you think of my revised language in the documentation in doc.go and log.go?