@peterbourgon 2017-03-06
Yes, it's great. I like that the previously floating comment about stack frame details now has a home on the unexported context type.